feat: unify block user
