+++ kit/dapp/src/lib/mutations/block-user/block-user-schema.ts
@@ -3,14 +3,16 @@ import { type ZodInfer, z } from '@/lib/utils/zod';
 /**
  * Zod schema for validating block user mutation inputs
  *
- * @property {string} address - The bond contract address
+ * @property {string} address - The contract address
  * @property {string} pincode - The pincode for signing the transaction
  * @property {string} account - The account to block
+ * @property {string} assetType - The type of asset
  */
 export const BlockUserSchema = z.object({
   address: z.address(),
   pincode: z.pincode(),
   account: z.address(),
+  assettype: z.assetType(),
 });
 
 export type BlockUserInput = ZodInfer<typeof BlockUserSchema>;
